Exported IP function stack
