+++ pyscf/pbc/cc/kccsd_rhf.py
@@ -31,7 +31,7 @@ from pyscf.pbc import scf
 from pyscf.pbc.mp.kmp2 import (get_frozen_mask, get_nocc, get_nmo,
                                padded_mo_coeff, padding_k_idx)
 from pyscf.pbc.cc import kintermediates_rhf as imdk
-from pyscf.pbc.cc.eom_kccsd_rhf_ip import ip_amplitudes_to_vector, ip_vector_to_amplitudes
+from pyscf.pbc.cc.eom_kccsd_rhf_ip import ipccsd
 from pyscf.pbc.cc.eom_kccsd_rhf_ea import ea_amplitudes_to_vector, ea_vector_to_amplitudes
 from pyscf.lib.parameters import LOOSE_ZERO_TOL, LARGE_DENOM
 from pyscf.lib import linalg_helper
@@ -360,6 +360,7 @@ class RCCSD(pyscf.cc.ccsd.CCSD):
                     'made_ip_imds', 'made_ea_imds', 'ip_partition',
                     'ea_partition', 'max_space', 'direct'])
         self._keys = self._keys.union(keys)
+        self.__imds__ = None
 
     @property
     def nkpts(self):
@@ -471,254 +472,13 @@ class RCCSD(pyscf.cc.ccsd.CCSD):
     def ao2mo(self, mo_coeff=None):
         return _ERIS(self, mo_coeff)
 
-    def vector_size_ip(self):
-        nocc = self.nocc
-        nvir = self.nmo - nocc
-        nkpts = self.nkpts
+    ipccsd = ipccsd
 
-        size = nocc + nkpts ** 2 * nocc ** 2 * nvir
-        return size
-
-    def ipccsd(self, nroots=1, koopmans=False, guess=None, partition=None,
-               kptlist=None):
-        '''Calculate (N-1)-electron charged excitations via IP-EOM-CCSD.
-
-        Kwargs:
-            nroots : int
-                Number of roots (eigenvalues) requested per k-point
-            koopmans : bool
-                Calculate Koopmans'-like (quasiparticle) excitations only, targeting via
-                overlap.
-            guess : list of ndarray
-                List of guess vectors to use for targeting via overlap.
-            partition : bool or str
-                Use a matrix-partitioning for the doubles-doubles block.
-                Can be None, 'mp' (Moller-Plesset, i.e. orbital energies on the diagonal),
-                or 'full' (full diagonal elements).
-            kptlist : list
-                List of k-point indices for which eigenvalues are requested.
-        '''
-        cput0 = (time.clock(), time.time())
-        log = logger.Logger(self.stdout, self.verbose)
-        nocc = self.nocc
-        nvir = self.nmo - nocc
-        nkpts = self.nkpts
-        if kptlist is None:
-            kptlist = range(nkpts)
-        size = self.vector_size_ip()
-        for k, kshift in enumerate(kptlist):
-            nfrozen = np.sum(self.mask_frozen_ip(np.zeros(size, dtype=int), kshift, const=1))
-            nroots = min(nroots, size - nfrozen)
-        if partition:
-            partition = partition.lower()
-            assert partition in ['mp', 'full']
-        self.ip_partition = partition
-        evals = np.zeros((len(kptlist), nroots), np.float)
-        evecs = np.zeros((len(kptlist), nroots, size), np.complex)
-
-        for k, kshift in enumerate(kptlist):
-            adiag = self.ipccsd_diag(kshift)
-            adiag = self.mask_frozen_ip(adiag, kshift, const=LARGE_DENOM)
-            if partition == 'full':
-                self._ipccsd_diag_matrix2 = ip_vector_to_amplitudes(self, adiag)[1]
-
-            if guess is not None:
-                guess_k = guess[k]
-                assert len(guess_k) == nroots
-                for g in guess_k:
-                    assert g.size == size
-            else:
-                guess_k = []
-                if koopmans:
-                    # Get location of padded elements in occupied and virtual space
-                    nonzero_opadding = padding_k_idx(self, kind="split")[0][kshift]
-
-                    for n in nonzero_opadding[::-1][:nroots]:
-                        g = np.zeros(size)
-                        g[n] = 1.0
-                        g = self.mask_frozen_ip(g, kshift, const=0.0)
-                        guess_k.append(g)
-                else:
-                    idx = adiag.argsort()[:nroots]
-                    for i in idx:
-                        g = np.zeros(size)
-                        g[i] = 1.0
-                        g = self.mask_frozen_ip(g, kshift, const=0.0)
-                        guess_k.append(g)
-
-            def precond(r, e0, x0):
-                return r / (e0 - adiag + 1e-12)
-
-            eig = linalg_helper.eig
-            if guess is not None or koopmans:
-                def pickeig(w, v, nr, envs):
-                    x0 = linalg_helper._gen_x0(envs['v'], envs['xs'])
-                    idx = np.argmax(np.abs(np.dot(np.array(guess_k).conj(), np.array(x0).T)), axis=1)
-                    return lib.linalg_helper._eigs_cmplx2real(w, v, idx)
-
-                evals_k, evecs_k = eig(lambda _arg: self.ipccsd_matvec(_arg, kshift), guess_k, precond, pick=pickeig,
-                                       tol=self.conv_tol, max_cycle=self.max_cycle,
-                                       max_space=self.max_space, nroots=nroots, verbose=self.verbose)
-            else:
-                evals_k, evecs_k = eig(lambda _arg: self.ipccsd_matvec(_arg, kshift), guess_k, precond,
-                                       tol=self.conv_tol, max_cycle=self.max_cycle,
-                                       max_space=self.max_space, nroots=nroots, verbose=self.verbose)
-
-            evals_k = evals_k.real
-            evals[k] = evals_k
-            evecs[k] = evecs_k
-
-            if nroots == 1:
-                evals_k, evecs_k = [evals_k], [evecs_k]
-
-            for n, en, vn in zip(range(nroots), evals_k, evecs_k):
-                r1, r2 = ip_vector_to_amplitudes(self, vn)
-                qp_weight = np.linalg.norm(r1) ** 2
-                logger.info(self, 'EOM root %d E = %.16g  qpwt = %0.6g',
-                            n, en, qp_weight)
-        log.timer('EOM-CCSD', *cput0)
-        self.eip = evals
-        return self.eip, evecs
-
-    def ipccsd_matvec(self, vector, kshift):
-        '''2ph operators are of the form s_{ij}^{ b}, i.e. 'jb' indices are coupled.'''
-        # Ref: Nooijen and Snijders, J. Chem. Phys. 102, 1681 (1995) Eqs.(8)-(9)
-        if not hasattr(self, 'imds'):
-            self.imds = _IMDS(self)
-        if not self.imds.made_ip_imds:
-            self.imds.make_ip(self.ip_partition)
-        imds = self.imds
-
-        vector = self.mask_frozen_ip(vector, kshift, const=0.0)
-        r1, r2 = ip_vector_to_amplitudes(self, vector)
-
-        t1, t2 = self.t1, self.t2
-        nkpts = self.nkpts
-        kconserv = self.khelper.kconserv
-
-        # 1h-1h block
-        Hr1 = -einsum('ki,k->i', imds.Loo[kshift], r1)
-        # 1h-2h1p block
-        for kl in range(nkpts):
-            Hr1 += 2. * einsum('ld,ild->i', imds.Fov[kl], r2[kshift, kl])
-            Hr1 += -einsum('ld,lid->i', imds.Fov[kl], r2[kl, kshift])
-            for kk in range(nkpts):
-                kd = kconserv[kk, kshift, kl]
-                Hr1 += -2. * einsum('klid,kld->i', imds.Wooov[kk, kl, kshift], r2[kk, kl])
-                Hr1 += einsum('lkid,kld->i', imds.Wooov[kl, kk, kshift], r2[kk, kl])
-
-        Hr2 = np.zeros(r2.shape, dtype=np.common_type(imds.Wovoo[0, 0, 0], r1))
-        # 2h1p-1h block
-        for ki in range(nkpts):
-            for kj in range(nkpts):
-                kb = kconserv[ki, kshift, kj]
-                Hr2[ki, kj] -= einsum('kbij,k->ijb', imds.Wovoo[kshift, kb, ki], r1)
-        # 2h1p-2h1p block
-        if self.ip_partition == 'mp':
-            nkpts, nocc, nvir = self.t1.shape
-            fock = self.eris.fock
-            foo = fock[:, :nocc, :nocc]
-            fvv = fock[:, nocc:, nocc:]
-            for ki in range(nkpts):
-                for kj in range(nkpts):
-                    kb = kconserv[ki, kshift, kj]
-                    Hr2[ki, kj] += einsum('bd,ijd->ijb', fvv[kb], r2[ki, kj])
-                    Hr2[ki, kj] -= einsum('li,ljb->ijb', foo[ki], r2[ki, kj])
-                    Hr2[ki, kj] -= einsum('lj,ilb->ijb', foo[kj], r2[ki, kj])
-        elif self.ip_partition == 'full':
-            Hr2 += self._ipccsd_diag_matrix2 * r2
-        else:
-            for ki in range(nkpts):
-                for kj in range(nkpts):
-                    kb = kconserv[ki, kshift, kj]
-                    Hr2[ki, kj] += einsum('bd,ijd->ijb', imds.Lvv[kb], r2[ki, kj])
-                    Hr2[ki, kj] -= einsum('li,ljb->ijb', imds.Loo[ki], r2[ki, kj])
-                    Hr2[ki, kj] -= einsum('lj,ilb->ijb', imds.Loo[kj], r2[ki, kj])
-                    for kl in range(nkpts):
-                        kk = kconserv[ki, kl, kj]
-                        Hr2[ki, kj] += einsum('klij,klb->ijb', imds.Woooo[kk, kl, ki], r2[kk, kl])
-                        kd = kconserv[kl, kj, kb]
-                        Hr2[ki, kj] += 2. * einsum('lbdj,ild->ijb', imds.Wovvo[kl, kb, kd], r2[ki, kl])
-                        Hr2[ki, kj] += -einsum('lbdj,lid->ijb', imds.Wovvo[kl, kb, kd], r2[kl, ki])
-                        Hr2[ki, kj] += -einsum('lbjd,ild->ijb', imds.Wovov[kl, kb, kj], r2[ki, kl])  # typo in Ref
-                        kd = kconserv[kl, ki, kb]
-                        Hr2[ki, kj] += -einsum('lbid,ljd->ijb', imds.Wovov[kl, kb, ki], r2[kl, kj])
-            tmp = (2. * einsum('xyklcd,xykld->c', imds.Woovv[:, :, kshift], r2[:, :])
-                      - einsum('yxlkcd,xykld->c', imds.Woovv[:, :, kshift], r2[:, :]))
-            Hr2[:, :] += -einsum('c,xyijcb->xyijb', tmp, t2[:, :, kshift])
-
-        return self.mask_frozen_ip(ip_amplitudes_to_vector(self, Hr1, Hr2), kshift, const=0.0)
-
-    def ipccsd_diag(self, kshift):
-        if not hasattr(self, 'imds'):
-            self.imds = _IMDS(self)
-        if not self.imds.made_ip_imds:
-            self.imds.make_ip(self.ip_partition)
-        imds = self.imds
-
-        t1, t2 = self.t1, self.t2
-        nkpts, nocc, nvir = t1.shape
-        kconserv = self.khelper.kconserv
-
-        Hr1 = -np.diag(imds.Loo[kshift])
-
-        Hr2 = np.zeros((nkpts, nkpts, nocc, nocc, nvir), dtype=t1.dtype)
-        if self.ip_partition == 'mp':
-            foo = self.eris.fock[:, :nocc, :nocc]
-            fvv = self.eris.fock[:, nocc:, nocc:]
-            for ki in range(nkpts):
-                for kj in range(nkpts):
-                    kb = kconserv[ki, kshift, kj]
-                    Hr2[ki, kj] = fvv[kb].diagonal()
-                    Hr2[ki, kj] -= foo[ki].diagonal()[:, None, None]
-                    Hr2[ki, kj] -= foo[kj].diagonal()[:, None]
-        else:
-            idx = np.arange(nocc)
-            for ki in range(nkpts):
-                for kj in range(nkpts):
-                    kb = kconserv[ki, kshift, kj]
-                    Hr2[ki, kj] = imds.Lvv[kb].diagonal()
-                    Hr2[ki, kj] -= imds.Loo[ki].diagonal()[:, None, None]
-                    Hr2[ki, kj] -= imds.Loo[kj].diagonal()[:, None]
-
-                    if ki == kconserv[ki, kj, kj]:
-                        Hr2[ki, kj] += np.einsum('ijij->ij', imds.Woooo[ki, kj, ki])[:, :, None]
-
-                    Hr2[ki, kj] -= np.einsum('jbjb->jb', imds.Wovov[kj, kb, kj])
-
-                    Wovvo = np.einsum('jbbj->jb', imds.Wovvo[kj, kb, kb])
-                    Hr2[ki, kj] += 2. * Wovvo
-                    if ki == kj:  # and i == j
-                        Hr2[ki, ki, idx, idx] -= Wovvo
-
-                    Hr2[ki, kj] -= np.einsum('ibib->ib', imds.Wovov[ki, kb, ki])[:, None, :]
-
-                    kd = kconserv[kj, kshift, ki]
-                    Hr2[ki, kj] -= 2. * np.einsum('ijcb,jibc->ijb', t2[ki, kj, kshift], imds.Woovv[kj, ki, kd])
-                    Hr2[ki, kj] += np.einsum('ijcb,ijbc->ijb', t2[ki, kj, kshift], imds.Woovv[ki, kj, kd])
-
-        return ip_amplitudes_to_vector(self, Hr1, Hr2)
-
-    def mask_frozen_ip(self, vector, kshift, const=LARGE_DENOM):
-        '''Replaces all frozen orbital indices of `vector` with the value `const`.'''
-        r1, r2 = ip_vector_to_amplitudes(self, vector)
-        nkpts, nocc, nvir = self.t1.shape
-        kconserv = self.khelper.kconserv
-
-        # Get location of padded elements in occupied and virtual space
-        nonzero_opadding, nonzero_vpadding = padding_k_idx(self, kind="split")
-
-        new_r1 = const * np.ones_like(r1)
-        new_r2 = const * np.ones_like(r2)
-
-        new_r1[nonzero_opadding[kshift]] = r1[nonzero_opadding[kshift]]
-        for ki in range(nkpts):
-            for kj in range(nkpts):
-                kb = kconserv[ki, kshift, kj]
-                idx = np.ix_([ki], [kj], nonzero_opadding[ki], nonzero_opadding[kj], nonzero_vpadding[kb])
-                new_r2[idx] = r2[idx]
-
-        return ip_amplitudes_to_vector(self, new_r1, new_r2)
+    @property
+    def imds(self):
+        if self.__imds__ is None:
+            self.__imds__ = _IMDS(self)
+        return self.__imds__
 
     def vector_size_ea(self):
         nocc = self.nocc
